Fix the Travis-CI Matrix job to have expansion for tox targets
The trick of using `env` to get a matrix expansion does not appear
to work when used within a matrix job, meaning we ended up with
only one tox target executed.

Here we define the Python job at the top-level, with the `env` key
for matrix expansion, and define the additional special jobs in the
`include` key, with empty `install` and `env` keys to mute them.

This is inspired from the travis-build project [1].

[1] https://github.com/travis-ci/travis-build/blob/master/.travis.yml