Create MySQL database on Toolforge shared server
Framadate requires a MySQL database [1].

For Wudele, let's use the shared `tools.db.svc.eqiad.wmflabs`
server and create a user database there.
The crendentials to access the server are available in
the automatically generated file `replica.my.cnf` [2].

Let's use Ansible to created the database on Toolforge,
using the `mysql_db` module [3]. Credentials-wise, it
can directly read the crendentials from the `cnf` config file.

However, on Toolforge, "Database names must start with the name of the
credential user", so we must somehow look it up from the config file:
we do so (ab)using the `setup` module [4] to read the config file as Ansible facts.
Since that module only works on `x.fact` files, we first alias the file
as `database.fact` with a symlink.

[1] <https://framacloud.org/fr/cultiver-son-jardin/framadate.html#mysql>
[2] <https://wikitech.wikimedia.org/wiki/Help:Toolforge/Database#User_databases>
[3] <https://docs.ansible.com/ansible/latest/modules/mysql_db_module.html>
[4] <https://docs.ansible.com/ansible/latest/modules/setup_module.html>