Move around the definition of `ansible_python_interpreter`
We rely on `ansible_python_interpreter` to use the local virtual environment
for the Ansible tasks, except for the bootstrapping tasks.

In 1d14cdc2922 we made the choice to define the override as
playbook variable of `main.yml`, and let the implicit `auto` for
the `bootstrap.yml` playbook.

This has the inconvenient side-effect of also applying for local actions
(when using `delegate_to: localhost`) which just errors out.

Here we define the override as host variables (in `hosts` file),
and cancel the override (using `auto`) in the `bootstrap.yml` playbook.