Completely disable email in Wudele
Emails are used to
 - send the poll creator two emails on poll creation
    - one to forward to participants
    - one with the admin link

 - remind a user of the polls he has

Arguably these are nice-to-have features, but if
SMTP is enabled, then Framadate will make it mandatory:
every poll will require an email address to be created.

This is a bit of a minefield in the Toolforge environment,
and may well be in violation of the Terms of Use [1].

Let's just disable email altogether for now.

[1] https://wikitech.wikimedia.org/wiki/Wikitech:Cloud_Services_Terms_of_use