Strip away URL parameters with Lighttpd URL rewriting
Spurious URL parameters confuse Wudele, resulting in broken links.

This is particularly problematic when shared on Facebook,
which appends a `fbclid` to all outgoing links.

Rather than with Framadate itself, the issue lies with
the URL rewriting for clean URLs (added in 4be80a60).

This makes sure that any additional URL parameters to a
clean URL is discarded/ignored in the rewriting output.

Bug: T254441